Avoid circular dependency between "feature" and "sfmData"
Move code from "feature" to "matching"
