+++ src/aliceVision/matching/CMakeLists.txt
@@ -16,6 +16,8 @@ set(matching_files_headers
   RegionsMatcher.hpp
   pairwiseAdjacencyDisplay.hpp
   supportEstimation.hpp
+  matchesFiltering.hpp
+  svgVisualization.hpp
 )
 
 # Sources
@@ -25,6 +27,8 @@ set(matching_files_sources
   matcherType.cpp
   RegionsMatcher.cpp
   supportEstimation.cpp
+  matchesFiltering.cpp
+  svgVisualization.cpp
 )
 
 alicevision_add_library(aliceVision_matching
