+++ src/aliceVision/matchingImageCollection/GeometricFilterMatrix_HGrowing.cpp
@@ -4,7 +4,7 @@
 // v. 2.0. If a copy of the MPL was not distributed with this file,
 // You can obtain one at https://mozilla.org/MPL/2.0/.
 
-#include <aliceVision/feature/svgVisualization.hpp>
+#include <aliceVision/matching/svgVisualization.hpp>
 #include "GeometricFilterMatrix_HGrowing.hpp"
 
 namespace aliceVision {
