+++ src/aliceVision/sfmData/View.hpp
@@ -111,6 +111,15 @@ public:
   }
 
   /**
+   * @brief Get view image height
+   * @return image height
+   */
+  std::pair<std::size_t, std::size_t> getImgSize() const
+  {
+    return {_width, _height};
+  }
+
+  /**
    * @brief Get the view id
    * @return view id
    */
