+++ src/samples/robustFundamentalF10/main_robustFundamentalF10.cpp
@@ -16,7 +16,7 @@
 #include <aliceVision/robustEstimation/conditioning.hpp>
 #include <aliceVision/robustEstimation/ACRansac.hpp>
 #include <aliceVision/robustEstimation/IRansacKernel.hpp>
-#include <aliceVision/feature/svgVisualization.hpp>
+#include <aliceVision/matching/svgVisualization.hpp>
 
 #include "dependencies/vectorGraphics/svgDrawer.hpp"
